# api/openapi.yaml
openapi: 3.1.0
info:
  title: Tabby Config Sync API
  version: 1.0.0
  description: Minimal API compatible with the Tabby desktop config sync client.
  license:
    name: All rights reserved
    identifier: LicenseRef-Proprietary
servers:
  - url: /
    description: Current host
security:
  - bearerAuth: []
paths:
  /api/1/user:
    get:
      operationId: getCurrentUser
      summary: Validate the sync token
      responses:
        "200":
          description: Authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/1/configs:
    get:
      operationId: listConfigs
      summary: List configs
      responses:
        "200":
          description: Config array
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Config"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createConfig
      summary: Create a config
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConfig"
      responses:
        "201":
          description: Created config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Config"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/1/configs/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
    get:
      operationId: getConfig
      summary: Get a config
      responses:
        "200":
          description: Config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Config"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateConfig
      summary: Update a config
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateConfig"
      responses:
        "200":
          description: Updated config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Config"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteConfig
      summary: Delete a config
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    User:
      type: object
      required: [id, username]
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
    Config:
      type: object
      required:
        - id
        - name
        - content
        - last_used_with_version
        - created_at
        - modified_at
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          maxLength: 255
        content:
          type: string
          description: Tabby YAML configuration stored verbatim.
        last_used_with_version:
          type: [string, "null"]
          maxLength: 32
        created_at:
          type: string
          format: date-time
        modified_at:
          type: string
          format: date-time
    CreateConfig:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          maxLength: 255
        content:
          type: string
          default: "{}"
        last_used_with_version:
          type: [string, "null"]
          maxLength: 32
    UpdateConfig:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name:
          type: string
          maxLength: 255
        content:
          type: string
        last_used_with_version:
          type: [string, "null"]
          maxLength: 32
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Invalid or missing bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Config not found for the authenticated user
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"